core: transfer views to same workspace as on destroyed output
Currently, when an output is destroyed, views are transferred to
the new output in roughly the same position as they were on the
old output, but they are all transferred to the first workspace.

This is annoying: if they were not on the first workspace, then
the user must have moved them, and after they're transferred, the
user must will have to move them agein.

Instead, in the (common?) case that the old and the new output
have the same number of workspaces, move each view to the same
workspace it was on before.